test: add tests for generating JSON for Sorenson
* Fixes the generate_json_for_encoding function (incorrect JSON was
  generated and that cause a lot of suffering).

* Adds tests for utils, to make sure this won't happen again (closes
  #12).

* Splits tests into separate files.
