+++ .travis.yml
@@ -40,7 +40,6 @@ env:
 
 python:
   - "2.7"
-  - "3.3"
   - "3.4"
   - "3.5"
