+++ cds_sorenson/utils.py
@@ -26,8 +26,6 @@
 
 from __future__ import absolute_import, print_function
 
-from os.path import dirname, normpath
-
 import requests
 from flask import current_app
